g jest teraz liczbą naturalną (tym razem naprawdę), poprawa ustawienia domyślnego arg_x, arg_y na arg_n/2, arg_m/2, zwiększenie minimalnych rozmiarów planszy (1x1 jest mało ciekawa: mrówka zderza się ze sobą, pojawiają się wtedy problemy z zaokrągleniem domyślnych arg_x, arg_y)

+++ mrowka.c
@@ -14,7 +14,7 @@ unsigned int liczba_mrowek;
 ///        2, jeśli wczytany plik ma nieprawidłowy format,
 ///        3, jeśli dane we wczytanym pliku przekraczają ograniczenia programu lub
 ///        0 w przeciwnym wypadku
-int inicjacja_siatki(int kolumny, int wiersze, double rand_perc, char * arg_input) {
+int inicjacja_siatki(int kolumny, int wiersze, int rand_perc, char * arg_input) {
     setlocale(LC_ALL, "");
     int i=-1, j=0;
     if (arg_input==NULL) {
